rename stripSanitizedParts
