+++ js/utils/stripSanitizedParts.js
